dfvm: added porosity variation to a code

+++ diffusion/math/Props.cpp
@@ -26,7 +26,7 @@
 #include <vector>
 
 
-Props::Props(const std::map<std::string, std::variant<int, double>> &params)
+Props::Props(const std::map<std::string, std::variant<double, int>> &params)
 
         : _params(params) {}
