+++ diffusion/math/Props.h
@@ -34,12 +34,11 @@ class Props {
 
 public:
     // ToDo: class Props as map, calcD to Convective
-    explicit Props(
-            const std::map<std::string, std::variant<int, double>> &params);
+    explicit Props(const std::map<std::string, std::variant<double, int>> &params);
 
     virtual ~Props() {}
 
-    std::map<std::string, std::variant<int, double>> _params;
+    std::map<std::string, std::variant<double, int>> _params;
 
     double _timePeriod;
     double _timeStep;
